src/em2mem/embedding/VLM2Vec/requirements.txt
torch
numpy==1.26.4
torchvision
transformers==4.52.3
huggingface-hub
peft
accelerate
datasets
scipy
tqdm
wandb
pillow
wrapt
py-cpuinfo
hjson
scikit-learn
scikit-image
qwen_vl_utils
ray
timm
sentencepiece
opencv-python
decord
hnswlib
einops
ninja
opencv-contrib-python
flash-attn  # check out orginal repo README for accelerating installation: https://github.com/Dao-AILab/flash-attention
qwen-vl-utils[decord]==0.0.8  # https://huggingface.co/Qwen/Qwen2.5-VL-3B-Instruct/discussions/30